% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{critical_power_model_3p}
\alias{critical_power_model_3p}
\title{3-Parameter Critical Power Model}
\usage{
critical_power_model_3p(t, cp, Wprime, Pmax)
}
\arguments{
\item{t}{Numeric. Running duration in seconds. Must be positive.}

\item{cp}{Numeric. Critical power in w. Must be positive.}

\item{Wprime}{Numeric. Anaerobic energy reserve in J. Must be positive.}

\item{Pmax}{Numeric. Maximal instantaneous power in w. Must be positive and greater than cp.}
}
\value{
Numeric. Power output in watts.
}
\description{
This function calculates the power output based on the 3-Parameter Critical Power model.
}
\examples{
critical_power_model_3p(t = 300, cp = 300, Wprime = 20000, Pmax = 1000)

}
\references{
Morton, R. H. "A 3-Parameter Critical Power Model." Ergonomics 39, no. 4 (April 1996): 611–19. https://doi.org/10.1080/00140139608964484.
}
