% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{di_prampero_running}
\alias{di_prampero_running}
\title{Convert Running Velocity to Power}
\usage{
di_prampero_running(
  velocity,
  distance,
  C1 = 3.86,
  C2 = 0.4,
  C3 = 2,
  BMR = 1.2,
  BSA = 1.8,
  BM = 70,
  eff = 0.2
)
}
\arguments{
\item{velocity}{Numeric. Running velocity in m/s.}

\item{distance}{Numeric. Running distance in meters.}

\item{C1}{Numeric. Parameter of the equation. Default is 3.86.}

\item{C2}{Numeric. Parameter of the equation. Default is 0.4.}

\item{C3}{Numeric. Parameter of the equation. Default is 2.}

\item{BMR}{Numeric. Basal metabolic rate in W/kg. Default is 1.2.}

\item{BSA}{Numeric. Body surface area in m^2. Default is 1.8.}

\item{BM}{Numeric. Body mass in kg. Default is 70.}

\item{eff}{Numeric. Assumed mechanical efficiency. Default is 0.2.}
}
\value{
Numeric. Power output in watts.
}
\description{
This function calculates power output based on running velocity and distance,
using the di Prampero equation.
}
\examples{
di_prampero_running(velocity = 5, distance = 1000)
di_prampero_running(velocity = 6, distance = 5000, BM = 60, BSA = 1.7)

}
\references{
Di Prampero, P. E., and S. Fusi. “Il Costo Energetico Della Corsa. / Energy Cost of Running.”
Medicina Dello Sport; Rivista Di Fisiopatologia Dello Sport 54, no. 1 (2001): 43–50.
http://articles.sirc.ca/search.cfm?id=S-925527 https://search.ebscohost.com/login.aspx?direct=true&db=s3h&AN=SPHS-925527&lang=fr&site=ehost-live.
}
