% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elite_runners}
\alias{elite_runners}
\title{Elite Runners Performance Data}
\format{
A data frame with \link{number of rows} rows and 9 variables:
\describe{
\item{Athlete}{Name of the athlete}
\item{Distance}{Running distance in meters}
\item{h}{Hours component of the running time}
\item{m}{Minutes component of the running time}
\item{s}{Seconds component of the running time}
\item{Time}{Total running time in seconds}
\item{VO2}{Measured VO2 max of the athlete, when available}
\item{BM}{Body mass of the athlete in kg}
\item{height}{Height of the athlete in cm}
}
}
\source{
\url{https://www.worldathletics.org/}
}
\usage{
elite_runners
}
\description{
A dataset containing performance data for elite runners across various distances.
}
\keyword{datasets}
