% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inv_di_prampero_running}
\alias{inv_di_prampero_running}
\title{Inverse Di Prampero Running Function}
\usage{
inv_di_prampero_running(
  Power,
  t,
  C1 = 3.86,
  C2 = 0.4,
  C3 = 2,
  BMR = 1.2,
  BSA = 1.8,
  BM = 70,
  eff = 0.2
)
}
\arguments{
\item{Power}{Numeric. Power output in watts.}

\item{t}{Numeric. Running duration in seconds.}

\item{C1}{Numeric. Energy cost of running per unit distance. Default is 3.86 J/(kg*m).}

\item{C2}{Numeric. Coefficient related to air resistance. Default is 0.4.}

\item{C3}{Numeric. Coefficient related to acceleration. Default is 2.}

\item{BMR}{Numeric. Basal metabolic rate in W/kg. Default is 1.2.}

\item{BSA}{Numeric. Body surface area in m^2. Default is 1.8.}

\item{BM}{Numeric. Body mass in kg. Default is 70.}

\item{eff}{Numeric. Assumed mechanical efficiency. Default is 0.2.}
}
\value{
Numeric. Estimated distance covered in meters.
}
\description{
This function calculates the distance covered given a power output and duration,
using an inverse of the Di Prampero running equation.
}
\examples{
inv_di_prampero_running(Power = 300, t = 600)

}
\references{
Di Prampero, P. E., and S. Fusi. “Il Costo Energetico Della Corsa. / Energy Cost of Running.”
Medicina Dello Sport; Rivista Di Fisiopatologia Dello Sport 54, no. 1 (2001): 43–50.
http://articles.sirc.ca/search.cfm?id=S-925527 https://search.ebscohost.com/login.aspx?direct=true&db=s3h&AN=SPHS-925527&lang=fr&site=ehost-live.
}
