% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{Kennelly_1906_P}
\alias{Kennelly_1906_P}
\title{Kennelly Power Model}
\usage{
Kennelly_1906_P(
  Distance,
  Ck,
  C1 = 3.86,
  C2 = 0.4,
  C3 = 2,
  BMR = 1.2,
  BSA = 1.8,
  BM = 70,
  eff = 0.2
)
}
\arguments{
\item{Distance}{Numeric. Running distance in meters (must be positive).}

\item{Ck}{Numeric. Empirical coefficient determined from a set of performances (must be positive).}

\item{C1}{Numeric. Parameter of the equation. Default is 3.86.}

\item{C2}{Numeric. Parameter of the equation. Default is 0.4.}

\item{C3}{Numeric. Parameter of the equation. Default is 2.}

\item{BMR}{Numeric. Basal metabolic rate in W/kg. Default is 1.2.}

\item{BSA}{Numeric. Body surface area in m^2. Default is 1.8.}

\item{BM}{Numeric. Body mass in kg. Default is 70.}

\item{eff}{Numeric. Assumed mechanical efficiency. Default is 0.2.}
}
\value{
Numeric. Power output in watts.
}
\description{
Calculates running power based on distance using Kennelly's model and di Prampero's equation.
}
\examples{
Kennelly_1906_P(Distance = 1000, Ck = 10)
}
\references{
Kennelly, A. E. "An Approximate Law of Fatigue in the Speeds of Racing Animals."
Proceedings of the American Academy of Arts and Sciences 42, no. 15 (1906): 275–331.
https://doi.org/10.2307/20022230.
}
