% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{Kennelly_1906_V}
\alias{Kennelly_1906_V}
\title{Kennelly Velocity Model}
\usage{
Kennelly_1906_V(Distance, Ck)
}
\arguments{
\item{Distance}{Numeric. Running distance in meters (must be positive).}

\item{Ck}{Numeric. Empirical coefficient determined from a set of performances (must be positive).}
}
\value{
Numeric. Running velocity in m/s.
}
\description{
Calculates running velocity based on distance using Kennelly's model.
}
\examples{
Kennelly_1906_V(Distance = 1000, Ck = 10)
}
\references{
Kennelly, A. E. "An Approximate Law of Fatigue in the Speeds of Racing Animals."
Proceedings of the American Academy of Arts and Sciences 42, no. 15 (1906): 275–331.
https://doi.org/10.2307/20022230.
}
