% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{Lloyd}
\alias{Lloyd}
\title{Lloyd Model}
\usage{
Lloyd(Pmax, g, MAP, k, Time, BMR = 1.2)
}
\arguments{
\item{Pmax}{Numeric. Maximum instantaneous power in W/kg (must be positive).}

\item{g}{Numeric. Time constant of the decrease of anaerobic metabolism contribution (must be positive).}

\item{MAP}{Numeric. Maximal aerobic power in W/kg (must be positive).}

\item{k}{Numeric. Time delay in the rise of aerobic metabolism power contribution from resting level to maximum (must be positive).}

\item{Time}{Numeric. Running duration in seconds (must be positive).}

\item{BMR}{Numeric. Basal metabolic rate in W/kg. Default is 1.2.}
}
\value{
Numeric. Power output in W/kg.
}
\description{
Calculates power output based on time using Lloyd's model.
}
\examples{
Lloyd(Pmax = 25, g = 0.05, MAP = 25, k = 10, Time = 100)
}
\references{
Lloyd. "WORLD RUNNING RECORDS AS MAXIMAL PERFORMANCES-OXYGEN DEBT AND OTHER LIMITING FACTORS."
In Circulation Research, 20:I218. LIPPINCOTT WILLIAMS & WILKINS 227 EAST WASHINGTON SQ,
PHILADELPHIA, PA 19106, 1967.
}
