% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{model_roy}
\alias{model_roy}
\title{Roy and Joyner Running Speed Model}
\usage{
model_roy(S, V0, B, b, d, Distance)
}
\arguments{
\item{S}{Numeric. Index representing the aerobic metabolic slowdown or reduced contribution as running distance increases (m/s). Must be positive.}

\item{V0}{Numeric. Extrapolated speed at 100 m (m/s). Must be positive.}

\item{B}{Numeric. Index representing an anaerobic boost at 100 m (m/s). Must be positive.}

\item{b}{Numeric. Determines the shape of the anaerobic boost curve as a function of distance. Must be positive.}

\item{d}{Numeric. Rate of decline of the anaerobic boost with increasing distance (m). Must be positive.}

\item{Distance}{Numeric. Running distance in meters. Must be positive.}
}
\value{
Numeric. World record running speed in m/s.
}
\description{
This function calculates world record running speed based on the Roy and Joyner model.
}
\examples{
model_roy(S = 0.1, V0 = 12, B = 2, b = 1.5, d = 200, Distance = 1000)
}
\references{
Roy, Tuhin K., Michael J. Joyner, Jonathon W. Senefeld, Chad C. Wiggins, and Timothy W. Secomb.
"An Empirical Model for World Record Running Speeds with Distance, Age, and Sex: Anaerobic and Aerobic Contributions to Performance."
Journal of Applied Physiology (Bethesda, Md.: 1985), June 27, 2024. https://doi.org/10.1152/japplphysiol.00033.2024.
}
