% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{peronnet_thibault_model}
\alias{peronnet_thibault_model}
\title{Péronnet-Thibault Running Performance Model}
\usage{
peronnet_thibault_model(T_est, A, MAP, E, Tmap, BMR = 1.2)
}
\arguments{
\item{T_est}{Running duration in seconds}

\item{A}{Anaerobic capacity in J/kg}

\item{MAP}{Maximal aerobic power in W/kg}

\item{E}{Endurance in W/kg/log(time in s)}

\item{Tmap}{Running duration eliciting MAP in seconds}

\item{BMR}{Basal metabolic rate in W/kg}
}
\value{
Power output in W/kg
}
\description{
This function calculates the power output for a given running duration based on
the model proposed by Péronnet and Thibault (1989).
}
\examples{
# Example usage (replace with realistic values):
peronnet_thibault_model(T_est = 600, A = 1500, MAP = 25, E = -1.5, Tmap = 420, BMR = 1.2)

}
\references{
Péronnet, F., & Thibault, G. (1989). Mathematical analysis of running performance
and world running records. Journal of applied physiology, 67(1), 453-465.
}
