% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{Riegel_time}
\alias{Riegel_time}
\title{Riegel Time Model}
\usage{
Riegel_time(Distance, a, b)
}
\arguments{
\item{Distance}{Numeric. Running distance in meters (must be positive).}

\item{a}{Numeric. Empirical constant dependent on measurement units (must be positive).}

\item{b}{Numeric. "Fatigue factor" constant as described by Riegel (must be positive).}
}
\value{
Numeric. Running time in seconds.
}
\description{
Calculates running time based on distance using Riegel's model.
}
\examples{
Riegel_time(Distance = 1000, a = 10, b = 1.06)
}
\references{
Riegel, Peter S. "Athletic Records and Human Endurance: A Time-vs.-Distance Equation
Describing World-Record Performances May Be Used to Compare the Relative Endurance
Capabilities of Various Groups of People." American Scientist 69, no. 3 (1981): 285–90.
http://www.jstor.org/stable/27850427.
}
