% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{Rumball_power}
\alias{Rumball_power}
\title{Rumball Power Model}
\usage{
Rumball_power(
  Distance,
  f,
  c,
  C1 = 3.86,
  C2 = 0.4,
  C3 = 2,
  BMR = 1.2,
  BSA = 1.8,
  BM = 70,
  eff = 0.2
)
}
\arguments{
\item{Distance}{Numeric. Running distance in meters (must be positive).}

\item{f}{Numeric. Empirical coefficient (slope of the relationship) referred to as the "fatigue coefficient" by Rumball (must be positive).}

\item{c}{Numeric. Empirical coefficient (intercept of the relationship).}

\item{C1}{Numeric. Parameter of the equation. Default is 3.86.}

\item{C2}{Numeric. Parameter of the equation. Default is 0.4.}

\item{C3}{Numeric. Parameter of the equation. Default is 2.}

\item{BMR}{Numeric. Basal metabolic rate in W/kg. Default is 1.2.}

\item{BSA}{Numeric. Body surface area in m^2. Default is 1.8.}

\item{BM}{Numeric. Body mass in kg. Default is 70.}

\item{eff}{Numeric. Assumed mechanical efficiency. Default is 0.2.}
}
\value{
Numeric. Power output in watts.
}
\description{
Calculates running power based on distance using Rumball's model and di Prampero's equation.
}
\examples{
Rumball_power(Distance = 1000, f = 0.5, c = 2)
}
\references{
Rumball, W. M., and C. E. Coleman. "Analysis of Running and the Prediction of Ultimate Performance."
Nature 228, no. 5267 (1970): 184–85. https://www.nature.com/articles/228184a0.
}
