% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{Rumball_velocity}
\alias{Rumball_velocity}
\title{Rumball Velocity Model}
\usage{
Rumball_velocity(Distance, f, c)
}
\arguments{
\item{Distance}{Numeric. Running distance in meters (must be positive).}

\item{f}{Numeric. Empirical coefficient (slope of the relationship) referred to as the "fatigue coefficient" by Rumball (must be positive).}

\item{c}{Numeric. Empirical coefficient (intercept of the relationship).}
}
\value{
Numeric. Running velocity in m/s.
}
\description{
Calculates running velocity based on distance using Rumball's model.
}
\examples{
Rumball_velocity(Distance = 1000, f = 0.5, c = 2)
}
\references{
Rumball, W. M., and C. E. Coleman. "Analysis of Running and the Prediction of Ultimate Performance."
Nature 228, no. 5267 (1970): 184–85. https://www.nature.com/articles/228184a0.
}
