% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wr_males}
\alias{wr_males}
\title{World Records for Male Athletes}
\format{
A data frame with \link{number of rows} rows and \link{number of columns} columns:
\describe{
\item{event}{Name of the event}
\item{time}{Record time}
\item{athlete}{Name of the athlete}
\item{country}{Athlete's country}
\item{date}{Date when the record was set}
...
}
}
\source{
\url{https://www.worldathletics.org/records/by-category/world-records}
}
\usage{
wr_males
}
\description{
A dataset containing world records for various events for male athletes.
}
\keyword{datasets}
