% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WR_males}
\alias{WR_males}
\title{World Records for Male Athletes}
\format{
A data frame with 1,232 rows and 3 columns:
\describe{
\item{Year}{The year of the world record, ranging from 1912 to 2023}
\item{Distance}{The running distance in meters. Possible values are 100, 200, 400, 800, 1500, 1609, 3000, 5000, 10000, 21100 (half marathon), and 42195 (marathon)}
\item{Time}{World record time in seconds for the given year and distance}
}
}
\source{
\url{https://www.worldathletics.org/records/by-category/world-records}
}
\usage{
WR_males
}
\description{
A dataset containing world record times for male athletes across 11 distances from 1912 to 2023 in long format.
}
\keyword{datasets}
